requirements.txt
asgiref==3.5.1
astroid==2.11.4
certifi==2020.12.5
charset-normalizer==2.0.12
colorama==0.4.4
coverage==5.3
dill==0.3.4
Django==4.0.4
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-ranged-response==0.2.0
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.8.6
pylint==2.13.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dotenv==0.15.0
pytz==2020.4
requests==2.27.1
sentry-sdk==1.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.5
wrapt==1.14.1
